Begin refactoring stake manager code.
This pulls out code from the old wstakemgr package (now part of udb)
that dealt with creating vote and revocation transactions.  The
transaction creation and notification processing is now performed
entirely by the wallet package and udb is only used for storage of
vote and revocation records.

Database handling has also been improved.  Votes and revocations are
no longer authored and signed during a DB update, and unique DB
updates are opened for publishing each of these created transactions.

Some stakepool-specific handling has also been removed since it should
no longer be required under the stakepoold architecture.

+++ rpc/legacyrpc/methods.go
@@ -549,7 +549,8 @@ func generateVote(icmd interface{}, w *wallet.Wallet) (interface{}, error) {
 		ExtendedBits: voteBitsExt,
 	}
 
-	ssgentx, err := w.GenerateVoteTx(blockHash, cmd.Height, ticketHash, voteBits)
+	ssgentx, err := w.GenerateVoteTx(blockHash, int32(cmd.Height), ticketHash,
+		voteBits)
 	if err != nil {
 		return nil, err
 	}
